Shared SetupFixture since both test projects requires to have TestContainers instance
