+++ .github/workflows/CodeQuality.yml
@@ -48,9 +48,6 @@ jobs:
       - name: Run Tests and Compute Coverage for NUnit - Hosting
         run: dotnet test Mercurio.Hosting.Tests/Mercurio.Hosting.Tests.csproj --no-restore --no-build --verbosity normal /p:CollectCoverage=true /p:CoverletOutput="../CoverageResults/" /p:MergeWith="../CoverageResults/coverage.json" /p:CoverletOutputFormat=\"opencover,json\"
 
-      - name: Run Tests and Compute Coverage for TUnit
-        run: dotnet test Mercurio.Tests.TUnit/Mercurio.Tests.TUnit.csproj --no-restore --no-build --verbosity normal /p:CollectCoverage=true /p:CoverletOutput="../CoverageResults/" /p:MergeWith="../CoverageResults/coverage.json" /p:CoverletOutputFormat=\"opencover,json\"
-
       - name: Sonarqube end
         env:
           SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
