+++ Mercurio.Core.Tests/RabbitMqContainerSetupFixture.cs
@@ -1,0 +1,50 @@
+﻿// -------------------------------------------------------------------------------------------------
+//  <copyright file="RabbitMqContainerSetupFixture.cs" company="Starion Group S.A.">
+// 
+//    Copyright 2025 Starion Group S.A.
+// 
+//    Licensed under the Apache License, Version 2.0 (the "License");
+//    you may not use this file except in compliance with the License.
+//    You may obtain a copy of the License at
+// 
+//        http://www.apache.org/licenses/LICENSE-2.0
+// 
+//    Unless required by applicable law or agreed to in writing, software
+//    distributed under the License is distributed on an "AS IS" BASIS,
+//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//    See the License for the specific language governing permissions and
+//    limitations under the License.
+// 
+//  </copyright>
+//  ------------------------------------------------------------------------------------------------
+
+namespace Mercurio.Tests
+{
+    using NUnit.Framework;
+
+    using Testcontainers.RabbitMq;
+
+    [SetUpFixture]
+    public class RabbitMqContainerSetupFixture
+    {
+        [OneTimeSetUp]
+        public async Task Setup()
+        {
+            RabbitMqContainer = new RabbitMqBuilder()
+                .WithImage("rabbitmq:4.2.0-alpine")
+                .WithUsername("guest")
+                .WithPassword("guest")
+                .Build();
+
+            await RabbitMqContainer.StartAsync();
+        }
+
+        public static RabbitMqContainer RabbitMqContainer { get; private set; }
+
+        [OneTimeTearDown]
+        public async Task Teardown()
+        {
+            await RabbitMqContainer.DisposeAsync();
+        }
+    }
+}
