+++ Mercurio.Tests/Messaging/RuntimeTracingTestFixture.cs
@@ -85,10 +85,8 @@ namespace Mercurio.Tests.Messaging
                 {
                     var connectionFactory = new ConnectionFactory
                     {
-                        HostName = "localhost",
-                        Port = 5672,
-                        UserName = "guest",
-                        Password = "guest"
+                        HostName = RabbitMqContainerSetupFixture.RabbitMqContainer.Hostname,
+                        Port = RabbitMqContainerSetupFixture.RabbitMqContainer.GetMappedPublicPort()
                     };
 
                     return connectionFactory;
@@ -97,10 +95,8 @@ namespace Mercurio.Tests.Messaging
                 {
                     var connectionFactory = new ConnectionFactory
                     {
-                        HostName = "localhost",
-                        Port = 5672,
-                        UserName = "guest",
-                        Password = "guest"
+                        HostName = RabbitMqContainerSetupFixture.RabbitMqContainer.Hostname,
+                        Port = RabbitMqContainerSetupFixture.RabbitMqContainer.GetMappedPublicPort()
                     };
 
                     return connectionFactory;
